refactor: use separated files do UserData interfaces and Loggers Enum
