+++ src/interfaces/enum-logger.ts
@@ -1,0 +1,6 @@
+export enum LogLevel {
+	INFO = 'info',
+	WARN = 'warn',
+	ERROR = 'error',
+	DEBUG = 'debug'
+}
