+++ src/interfaces/userdata-interface.ts
@@ -1,0 +1,3 @@
+export interface UserData {
+	[key: string]: string;
+}
